+++ packages/core/src/loader/DocumentLoader.tsx
@@ -7,15 +7,14 @@
  */
 
 import * as React from 'react';
-import { ProgressBar } from '../components/ProgressBar';
 import { Spinner } from '../components/Spinner';
 import { useIsMounted } from '../hooks/useIsMounted';
 import { TextDirection, ThemeContext } from '../theme/ThemeContext';
+import type { CharacterMap } from '../types/CharacterMap';
 import type { DocumentAskPasswordEvent, VerifyPassword } from '../types/DocumentAskPasswordEvent';
 import type { PdfJs } from '../types/PdfJs';
 import { classNames } from '../utils/classNames';
 import { PdfJsApi } from '../vendors/PdfJsApi';
-import { CharacterMap } from '../Viewer';
 import { AskForPasswordState, SubmitPassword } from './AskForPasswordState';
 import { AskingPassword } from './AskingPassword';
 import { CompletedState } from './CompletedState';